refactor(mobile): split ios pod manifest
